capstone/requirements.txt
validate_docbr
crispy-bootstrap4
django-crispy-forms